refactor : Storage 파일에 타입적용
