+++ README.md
@@ -14,10 +14,10 @@ FE 레벨1 영화 리뷰 미션
 
 **영화 상세정보 조회**
 
-- [ ] 영화 포스터나 제목 클릭시 정보를 보여주는 모달이 나타난다
-- [ ] 상세정보에서 내 별점을 매길 수 있다
-  - [ ] 별1개당 2점이 채워진다
-  - [ ] 새로고침해도 별점은 유지되어야한다
+- [x] 영화 포스터나 제목 클릭시 정보를 보여주는 모달이 나타난다
+- [x] 상세정보에서 내 별점을 매길 수 있다
+  - [x] 별1개당 2점이 채워진다
+  - [x] 새로고침해도 별점은 유지되어야한다
 
 **영화 검색 기능**
 
@@ -28,7 +28,7 @@ FE 레벨1 영화 리뷰 미션
 **UI/UX 개선하기**
 
 - [x] 디바이스의 너비에 따라 한 줄에 나타나는 영화 개수가 줄어든다.
-- [ ] 디바이스의 너비에 따라 모달의 배치가 변한다.
+- [x] 디바이스의 너비에 따라 모달의 배치가 변한다.
 
 **오류**
