--- src/storages/LocalStorage.js
@@ -1,15 +1,0 @@
-class LocalStorage {
-  static get(key) {
-    return JSON.parse(localStorage.getItem(key));
-  }
-
-  static set(key, value) {
-    localStorage.setItem(key, JSON.stringify(value));
-  }
-
-  static remove(key) {
-    localStorage.removeItem(key);
-  }
-}
-
-export default LocalStorage;
